requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bootstrap4==0.1.0
Django==3.0.7
django-bootstrap4==1.1.1
django-braces==1.14.0
django-extensions==2.2.9
django-model-utils==4.0.0
django-star-ratings==0.9.1
django-taggit==2.0.0
importlib-metadata==2.1.2
mysqlclient==1.4.6
Pillow==7.1.1
pytz==2019.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
swapper==1.1.2.post1
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
